Extract incremental downloads to a new module
