+++ lib/fetch.js
@@ -1,7 +1,8 @@
-var S3 = require('s3')
-var RSVP = require('rsvp')
-var path = require('path')
-var fs = require('fs')
+let S3 = require('s3')
+let RSVP = require('rsvp')
+let path = require('path')
+let fs = require('fs')
+let getNewDocsToIndex = require('./identify-docs-to-index')
 
 function mkdirp (dir) {
   return new RSVP.Promise(function (resolve, reject) {
@@ -82,55 +83,6 @@ function filterReleaseDocs (document) {
   var tag = key[key.length - 2]
   var versionRegex = /v\d+\.\d+\.\d+$/
   return versionRegex.test(tag) && /-docs\.json/.test(key)
-}
-
-function getNewDocsToIndex (db, docs) {
-  return db.allDocs({
-    keys: ['project-ember', 'project-ember-data'],
-    include_docs: true
-  }).then(projectVersions => {
-    let docsToDownload = docs
-
-    if (projectVersions.total_rows > 0) {
-      let [ {doc: emberProjectDoc}, {doc: emberDataProjectDoc} ] = projectVersions.rows
-
-      // Save now for merging later
-      let cachedDocsFolder = path.join('tmp', 'cached-docs')
-      require('mkdirp').sync(cachedDocsFolder)
-      fs.writeFileSync(path.join(cachedDocsFolder, 'project-ember.json'), JSON.stringify(emberProjectDoc))
-      fs.writeFileSync(path.join(cachedDocsFolder, 'project-ember-data.json'), JSON.stringify(emberDataProjectDoc))
-
-      let indexedEmberVersions = []
-      let indexedEmberDataVersions = []
-
-      emberProjectDoc.data.relationships['project-versions'].data.forEach(doc => {
-        indexedEmberVersions.push(doc.id.replace('ember-', ''))
-      })
-
-      emberDataProjectDoc.data.relationships['project-versions'].data.forEach(doc => {
-        indexedEmberDataVersions.push(doc.id.replace('ember-data-', ''))
-      })
-
-      docsToDownload = docsToDownload.filter(doc => {
-        let docString = doc.Key.split('/')
-        let projectName = docString[2].replace('-docs.json', '')
-        let projectVersion = docString[1].replace('v', '')
-
-        if (projectName === 'ember') {
-          return indexedEmberVersions.indexOf(projectVersion) === -1
-        } else if (projectName === 'ember-data') {
-          return indexedEmberDataVersions.indexOf(projectVersion) === -1
-        }
-      })
-
-      if (docsToDownload.length === 0) {
-        console.log('No more docs to index')
-        process.exit(0)
-      }
-    }
-
-    return RSVP.resolve(docsToDownload)
-  })
 }
 
 module.exports = function fetch (db, releaseToGenDocFor) {
